Adding a note
- Dear future Sebastian,
Hey man, I wanted to apologize because I just realized that the masses you are using are instance masses for the bus's current mass, which is not getting updated as the route executes. As you might understand, this is a problem. So, to remedy this, you are going to need to swap the bus mass instances with bus mass profiles. This, of course, further complicates having random chance for stops, but that could be done through an external ridership tool. Maybe a 'n_riders' column. Food for thought.
Anyways, best of luck!

-Past Sebastian