Failed attempt at solving oscillation issue
So, in my process for trying to solve the oscillation issue, I have rewritten maintain, (now on v3) and I have also re-adjusted stops to make sure the proper stopping is done. Part of the wild oscillations were due to strange issues where if the end of the route was not specified as a stop, it would break. For some reason, now, when the velocity is attempting to speed limit break, some weird conditions cause it to begin accelerating. May warrant rewriting brake v2.

Frustrating. But darn close. For some reason, it worked when the speed limit margins were wider.