Continued working on the Periodicity issue
Found an issue where, depending on period, some stops are not visited. This is causing issues with assigning ridership and stops, and I need to figure out a way to pass the stop ID with the ridership as a method of confirming ridership change. Or, leave it as is and get other aspects working. TBD based on discussion with dan.