Inconsistency Identified.
Upon further investigation, my suspicion about the points was partially correct. It indeed seems that there are a significant number of points in the route dataset I have. However, part of the reason for this is that both inbound and outbound routes are included in the same point series. My original assumption was incorrect. Unfortunately, without more intricate knowledge of the route, it is hard to extrapolate which of the route segments are used for inbound or outbound. I'm toying with the idea of using the asociated transit stop dataset to see if there's an overlap, but I can't be certain that will work. I was able to identify the excess of points through a clse examination of the ballard bridge area of route 17, an area I'm extensively familliar with. Perhaps the key lies in shape_length in the original dataset.