Loaded all data, got to filtering elevation noise.
Everything seems hunky-dory now, distances are lining up, and general elevation profiles seem accurate. The only thing I'm not quite certain of is how to filter out the noise of trees, as the DSM seems to be the most ideal. I think I've settled on a rolling median with a window of 7 rather than the savisky-golay filter, as the noise observed was not reduced regardless of parameters - and it was cutting off some rather critical features of the data.
Next Up:
- Continue step-by-step reworking of route_map.
- Find Bus Parameter Database
- Rework BusModel
- Rework TripDynamics to work with the current iteration of route_map.