Got the driving logic to work.
Turns out, importing the acceleration profile properly makes a significant difference. A potential alternative acceleration method has been coded, but I have yet to test it and as such am uncertain about using it. The profile generated appears reasonable with the current method in Physics_Engine, so I am going to leave it as-is. Gradually increasing the breaking factor works, and it has an impact on regenerative braking from my prelim tests!

Next on the list:
- Put the driving logic into its own file
- Get the script into its own file
- Get the objects into their own file
- Finalize Acceleration, braking, and maintenance tests
- Check if rest time is working
- Set up methodology for itertools and multiprocessing!