Trying to determine metrics and causes for discrepancy
Based on time alone, it seems the model is reasonable. More replicates are required, though. However, when it comes to accurately predicting the acceleration and velocity histograms, the model currently struggles in two areas - the first being the rests. The rests at stoplights are uniform, and the rests at signals are uniformely based on number of riders. However, the model, at present, does not account for stops signs or otherwise unmarked intersections. As an example, route 45 goes through a section on campus that has several crosswalks it stops for, without signage. In other words, the total time at rest will be higher, across the board. How much higher, we need more replicates to know. The second area is in the acceleration/velocity - that being, the velocity skews low. I've a running theory that the acceleration, particularly while braking, is to blame for this, as after fitting a new acceleration profile to the aggregate histogram of the collected data, the skew remains. I'm uncertain how to characterize this skew, nor how to remedy it, but unfortunately opening up the variable for bus deceleration did not do the trick.
From the presentation:
- Order figures from best, then show why the profile is not nessecarily a good figure.
- Use proper T-Test to illustrate time relevance
- More replicates
- I've built the model, I'm now showing it works reasonably well
- Then, once the model is shown to be working reasonably, we can move on to how it can be used (bins, runs en masse, statistics on loads)
- That should be good enough for first paper
- Once that's all said and done, we can get to using it to test on cell impact, voltage, etc.